pyproject.toml
[tool.poetry]
name = "dkc-api"
version = "1.0.0"
description = "Connector to DKC API v1"
authors = ["Alexander Dr. <sasna142@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.5.3"
requests = "^2.26.0"
pydantic = "^1.8.2"
pytz = "^2021.1"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
